requirements.txt
icalendar==3.9.*
requests_ntlm==0.2.*
requests==2.7.*
beautifulsoup4==4.4.*
lxml
pyconstraints
configparser
tabulate
vcrpy
pytest>=2.2
pytest-sugar>=0.5
pytest-cov>=2.2
